Add podcasts and some styling changes.
