+++ package.json
@@ -13,6 +13,7 @@
     "@testing-library/user-event": "^13.5.0",
     "mapbox-gl": "^3.1.2",
     "react": "^18.2.0",
+    "react-audio-player": "^0.17.0",
     "react-dom": "^18.2.0",
     "react-map-gl": "^7.1.7",
     "react-scripts": "5.0.1",
@@ -41,5 +42,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
   }
 }
